Move domain chain specs to appropriate modules
